+++ src/components/icure-form/fields/items-list-field/items-list-field.ts
@@ -1,10 +1,13 @@
 import { html, TemplateResult } from 'lit'
 import { Field } from '../../../common'
 import { property } from 'lit/decorators.js'
+import { Suggestion } from '../../../../generic'
 
 export class ItemsListField extends Field {
 	@property() multiline: boolean | string = false
 	@property() lines = 1
+	@property() suggestionProvider?: (terms: string[]) => Promise<Suggestion[]>
+	@property() linksProvider?: (sug: Suggestion) => Promise<{ href: string; title: string } | undefined>
 	override renderSync(): TemplateResult {
 		return html`<icure-text-field
 			schema="items-list"
@@ -15,6 +18,10 @@ export class ItemsListField extends Field {
 			.displayedLabels="${this.displayedLabels}"
 			.defaultLanguage="${this.defaultLanguage}"
 			.languages="${this.languages}"
+			?suggestions=${!!this.suggestionProvider}
+			?links=${!!this.linksProvider}
+			.suggestionProvider=${this.suggestionProvider}
+			.linksProvider=${this.linksProvider}
 			.ownersProvider=${this.ownersProvider}
 			.valueProvider=${this.valueProvider}
 			.validationErrorsProvider=${this.validationErrorsProvider}